requirements.txt
requests==2.9.1
pysparklines==0.9
configobj==5.0.6
boto3==1.2.3
